@@ -1,5 +1,5 @@
 import React from "react";
-import { BrowserRouter, Route, Switch } from "react-router-dom";
+import { BrowserRouter, Route, Routes } from "react-router-dom";
 import { FibonacciPage } from "../fibonacci-page/fibonacci-page";
 import { ListPage } from "../list-page/list-page";
 import { MainPage } from "../main-page/main-page";
@@ -14,29 +14,15 @@ function App() {
   return (
     <div className="app">
       <BrowserRouter>
-        <Switch>
-          <Route path="/" exact>
-            <MainPage />
-          </Route>
-          <Route path="/recursion">
-            <StringComponent />
-          </Route>
-          <Route path="/fibonacci">
-            <FibonacciPage />
-          </Route>
-          <Route path="/sorting">
-            <SortingPage />
-          </Route>
-          <Route path="/stack">
-            <StackPage />
-          </Route>
-          <Route path="/queue">
-            <QueuePage />
-          </Route>
-          <Route path="/list">
-            <ListPage />
-          </Route>
-        </Switch>
+        <Routes>
+          <Route path="/" element={<MainPage />}></Route>
+          <Route path="/recursion" element={<StringComponent />}></Route>
+          <Route path="/fibonacci" element={<FibonacciPage />}></Route>
+          <Route path="/sorting" element={<SortingPage />}></Route>
+          <Route path="/stack" element={<StackPage />}></Route>
+          <Route path="/queue" element={<QueuePage />}></Route>
+          <Route path="/list" element={<ListPage />}></Route>
+        </Routes>
       </BrowserRouter>
     </div>
   );
